Added Firebase image Storage
Added firebase image storage function
